requirements.txt
pytest
pytest_bdd
selenium==3.141.0
jsonpickle
pony
pymysql
mysql-connector-python-rf
urllib3==1.26.16
allure-pytest